requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0

# Web API and GUI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
requests>=2.31.0
pillow>=10.0.0
python-multipart>=0.0.6

# Optional: CUDA support (uncomment if using GPU)
# torch[gpu]>=2.0.0